+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [7.10.0-exp.6] - 2022-09-07
+- Added an extra "-render-graph-reuse-tests" command line option and define to the player so that RenderGraph test code can be enabled while still using the regular (non-RG) reference images.
+- Add support for LinuxHeadlessSimulation platform
+
 ## [7.10.0-exp.5] - 2022-08-30
 - When using the Test Scene Asset, if the "SceneName_RPAssetName" reference image is not found, fall back to the base "SceneName" reference image.